+++ src/app/app.module.ts
@@ -2,6 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Routes, RouterModule } from "@angular/router";
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
@@ -10,6 +11,7 @@ import { FiveDayForecastComponent } from './five-day-forecast/five-day-forecast.
 import { NavbarComponent } from './navbar/navbar.component';
 import { LocationService } from './services/location.service';
 import { NotFoundComponent } from './not-found/not-found.component';
+import { WeatherService } from './services/weather.service';
 
 // Routes Angular needs to know to access other "pages" using the navbar.
 const appRoutes: Routes = [
@@ -32,9 +34,10 @@ const appRoutes: Routes = [
   imports: [
     BrowserModule,
     FormsModule,
+    HttpClientModule,
     RouterModule.forRoot(appRoutes)
   ],
-  providers: [ LocationService ],
+  providers: [ LocationService, WeatherService ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }