+++ src/app/five-day-forecast/five-day-forecast.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { LocationService } from '../services/location.service'
+import { WeatherService } from '../services/weather.service'
 import { Subscription } from 'rxjs';
+import { env } from '../../environments/environment'
 
 @Component({
   selector: 'app-five-day-forecast',
@@ -12,15 +14,29 @@ export class FiveDayForecastComponent implements OnInit {
   private latitude: number;
   private longitude: number;
   private currentLocation: Subscription;
+  private forecasts: any;
+  private city: string;
 
-  constructor(private locationService: LocationService) {}
+  constructor(private _locationService: LocationService, 
+              private _weatherService: WeatherService) {}
+
+  getForecasts(forecasts){
+    this.forecasts = forecasts.list;
+    this.city = forecasts.city.name;
+  }
 
   ngOnInit(): void {
     // Retrieve the user's current location.
-    this.currentLocation = this.locationService.retrieveCurrentPosition().subscribe(position => {
+    this.currentLocation = this._locationService.retrieveCurrentPosition().subscribe(position => {
       this.latitude = position.coords.latitude;
       this.longitude = position.coords.longitude;
       //console.log(this.latitude, this.longitude);
+
+      // Get the forecasts' information from the above geographic coordinates.
+      this._weatherService.getForecastsByCoordinates(this.latitude, this.longitude).subscribe(response => {
+        //console.log(response);
+        this.getForecasts(response);
+      });
     });
   }