+++ src/environments/environment.prod.ts
@@ -1,6 +1,6 @@
-export const environment = {
+export const env = {
   production: true,
-  appID: 'bf4b4ba5c71e63e094c62cc9a10efa56',
+  appID: '&APPID=bf4b4ba5c71e63e094c62cc9a10efa56',
   baseURL: 'http://api.openweathermap.org/data/2.5/',
-  units: 'imperial'
+  units: '&units=imperial'
 };