+++ src/environments/environment.ts
@@ -2,9 +2,9 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-export const environment = {
+export const env = {
   production: false,
-  appID: 'bf4b4ba5c71e63e094c62cc9a10efa56',
+  appID: '&APPID=bf4b4ba5c71e63e094c62cc9a10efa56',
   baseURL: 'http://api.openweathermap.org/data/2.5/',
-  units: 'imperial'
+  units: '&units=imperial'
 };